% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisTopInteractors.R
\name{VisTopInteractors}
\alias{VisTopInteractors}
\title{A bar graph of the top n protein interactors.}
\usage{
VisTopInteractors(proteins, poi)
}
\value{
Returns a bar graph that visualizese the top n proteins generated from TopInteractors function
}
\description{
A function that utilize the Y2H-SCORES software functions (Banerjee et al., 2020) to find the top n interactors with the protein of interest.
to work.
}
\examples{
\dontrun{

CancerProtInteractors::VisTopInteractors( proteins = list_of_top_interacting_proteins <dataframe consist of prey, Borda_scores>,
                                          poi = name of protein of interest <chr>)
}

}
\references{

}
