% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y2HSCORES_output}
\alias{Y2HSCORES_output}
\title{Gene Counts From An RNAseq Experiment 1}
\format{
A matrix with columns:
\describe{
\item{Condition.1}{First condition, under which least light was provided
to plants.}
\item{Condition.2}{Second condition, under which most light was provided to plants.}
\item{Condition.3}{Third condition, the control.}
}
}
\source{
University of Guelph, Canada.
}
\usage{
Y2HSCORES_output
}
\description{
An RNAseq experiment conductd using bean plants from 2016 in Canada.
}
\examples{
\dontrun{
 Y2HSCORES_output
}
}
\keyword{datasets}
